requirements/base.txt
# 基础依赖
pytest>=7.4.3
requests>=2.31.0
playwright>=1.40.0
python-dotenv>=1.0.0
PyYAML>=6.0.1
pymysql>=1.1.0
redis>=5.0.1
cryptography>=41.0.3
openpyxl>=3.1.2